Conflicts:
	TightDB.vcxproj
	TightDB.vcxproj.filters
	src/tightdb/Makefile
	test/main.cpp

+++ doc/ref_cpp/examples/ex_query_dynamic_equal.cpp
@@ -1,6 +1,7 @@
 // @@Example: ex_cpp_dyn_query_equals @@
 // @@Fold@@
 #include <tightdb.hpp>
+#include <assert.h>
 
 using namespace tightdb;
 using namespace std;
@@ -38,7 +39,7 @@ int main()
     assert(view1.size() == 1 && view1.get_string(0,0) == "Frank");
 
     // Find rows where age (column 1) == 56
-    TableView view2 = table->where().equal(1, int64_t(56)).find_all();
+    TableView view2 = table->where().equal(1, 56).find_all();
     assert(view2.size() == 1 && view2.get_string(0,0) == "Frank");
 
     // Find rows where male (column 2) == true
